enable DefineBlock to format correctly
