+++ ts/block/block.ts
@@ -95,6 +95,7 @@ export abstract class Block extends HTMLElement {
                 target.parent = this;
                 child.parent = null;
                 child.kill();
+                target.parent.format();
                 this.get_root().format();
                 return true;
             }
@@ -110,8 +111,83 @@ export abstract class Block extends HTMLElement {
         this.determine_pos(this.get_x(), this.get_y(), this.determine_width());
     }
 
-    abstract determine_pos(x: number, y: number, res: FormatResult): void;
-    abstract determine_width(): FormatResult;
+    determine_pos(x: number, y: number, res: FormatResult) {
+        const center: number = x - res.x;
+        this.set_left(x);
+        this.set_top(y);
+        let offset: number = center + res.leftmost;
+        for (let i = 0; i < res.childrens.length; ++i) {
+            const child_area = (res.childrens[i].rightmost - res.childrens[i].leftmost);
+            this.get_children()[i].determine_pos(
+                offset + child_area * 0.5 + res.childrens[i].x,
+                y + this.get_height() + res.childrens[i].bottomdiff,
+                res.childrens[i]
+            );
+            offset += child_area;
+        }
+    }
+    determine_width(): FormatResult {
+        const children = this.get_children();
+        if (this.is_empty() || children.length == 0 || this.classList.contains("pyramid-block-folding")) {
+            this.style.minWidth = BlockConst.UNIT_WIDTH + "px";
+            return {
+                x: 0,
+                leftmost: -this.get_width() * 0.5,
+                rightmost: this.get_width() * 0.5,
+                bottommost: this.get_height(),
+                bottomdiff: 0,
+                childrens: [],
+            };
+        }
+        if (children.length == 1) {
+            const res = children[0].determine_width();
+            this.style.minWidth = BlockConst.UNIT_WIDTH + "px";
+            return {
+                x: res.x,
+                leftmost: res.leftmost,
+                rightmost: res.rightmost,
+                bottommost: res.bottommost + this.get_height(),
+                bottomdiff: res.bottomdiff,
+                childrens: [res],
+            };
+        }
+        let width = BlockConst.UNIT_WIDTH;
+        let leftmost = 0.0;
+        let rightmost = 0.0;
+        let bottommost = 0.0;
+        let bottomdiff = 0.0;
+        let childrens: FormatResult[] = [];
+        let i = 0;
+        for (const child of children) {
+            const res = child.determine_width();
+            leftmost += res.leftmost;
+            rightmost += res.rightmost;
+            if (i == 0) {
+                width += res.rightmost - res.x - child.get_width() * 0.5;
+            } else if (i == children.length - 1) {
+                width += res.x - child.get_width() * 0.5 - res.leftmost;
+            } else {
+                width += res.rightmost - res.leftmost;
+            }
+            bottommost = Math.max(bottommost, res.bottommost);
+            bottomdiff = Math.max(bottomdiff, bottomdiff + this.get_height() - BlockConst.UNIT_HEIGHT);
+            childrens.push(res);
+            i += 1;
+        }
+        this.style.minWidth = width + "px";
+        const x = leftmost
+            + (childrens[0].x - childrens[0].leftmost)
+            + (children[0].get_width() * 0.5 - BlockConst.UNIT_HALF_WIDTH)
+            + this.get_width() * 0.5;
+        return {
+            x: x,
+            leftmost: leftmost,
+            rightmost: rightmost,
+            bottommost: bottommost + this.get_height(),
+            bottomdiff: bottomdiff,
+            childrens: childrens,
+        };
+    }
 
     abstract is_empty(): boolean;
     abstract kill(): void;
