+++ ts/block/block_const.ts
@@ -3,4 +3,5 @@ export namespace BlockConst {
     export const UNIT_HALF_WIDTH: number = 50.0;
     export const UNIT_HEIGHT: number = 50.0;
     export const UNIT_HALF_HEIGHT: number = 25.0;
+    export const DEFINE_BLOCK_BORDER: number = 5;
 }
