+++ ts/block/concrete_block/define_block.ts
@@ -1,5 +1,6 @@
 import { TempPyramidType, TempPyramidTypeTree, TypeEnv } from "../inference/typeenv.js";
 
+import { Block } from "../block.js";
 import { BlockConst } from "../block_const.js";
 import { ParentBlock } from "../parent_block.js";
 import { TypedBlock } from "../typed_block.js";
@@ -45,6 +46,7 @@ export class DefineBlock extends ParentBlock {
                 break;
             }
         }
+        this.style.height = BlockConst.UNIT_WIDTH + BlockConst.DEFINE_BLOCK_BORDER + "px";
         this.set_content(content);
         this.appendChild(new EmptyBlock(this));
         this.appendChild(new EmptyBlock(this));
@@ -135,43 +137,39 @@ export class DefineBlock extends ParentBlock {
 
     override determine_pos(x: number, y: number, res: FormatResult) {
         const center: number = x - res.x;
+        let strech_width = res.childrens[0].rightmost + Math.abs(res.childrens[0].x);
+        this.style.minWidth = (strech_width * 2 + BlockConst.DEFINE_BLOCK_BORDER * 2) + "px";
+        this.style.minHeight = BlockConst.UNIT_HEIGHT + res.childrens[0].bottommost + BlockConst.DEFINE_BLOCK_BORDER * 2 + "px";
         this.set_left(x);
-        this.set_top(y);
+        this.set_top(y - BlockConst.DEFINE_BLOCK_BORDER);
+
         let offset: number = center + res.leftmost;
-        for (let i = 0; i < res.childrens.length; ++i) {
-            const child_area = (res.childrens[i].rightmost - res.childrens[i].leftmost);
-            this.get_children()[i].determine_pos(
-                offset + child_area * 0.5 + res.childrens[i].x,
-                y + this.get_height(),
-                res.childrens[i]
-            );
-            offset += child_area;
+        if (res.childrens.length !== 2) {
+            throw new Error("Pyramid Error: Wrong Define Block");
         }
+
+        // Function Wrapper Field
+
+        // Function Logic Parts
+        this.get_children()[0].determine_pos(
+            this.get_x(),
+            (y - (this.get_height() - BlockConst.UNIT_HEIGHT) * 0.5) + BlockConst.UNIT_HEIGHT + BlockConst.DEFINE_BLOCK_BORDER + res.childrens[0].bottomdiff,
+            res.childrens[0]
+        );
+        // Function Apply Parts
+        this.get_children()[1].determine_pos(
+            this.get_x(),
+            (y - (this.get_height() - BlockConst.UNIT_HEIGHT) * 0.5) + res.childrens[0].bottommost + BlockConst.UNIT_HEIGHT + BlockConst.DEFINE_BLOCK_BORDER * 2 + res.childrens[0].bottomdiff,
+            res.childrens[1]
+        );
     }
     override determine_width(): FormatResult {
         const children = this.get_children();
-        if (this.is_empty() || children.length == 0 || this.classList.contains("pyramid-block-folding")) {
-            this.style.minWidth = BlockConst.UNIT_WIDTH + "px";
-            return {
-                x: 0,
-                leftmost: -this.get_width() * 0.5,
-                rightmost: this.get_width() * 0.5,
-                childrens: [],
-            };
-        }
-        if (children.length == 1) {
-            const res = children[0].determine_width();
-            this.style.minWidth = BlockConst.UNIT_WIDTH + "px";
-            return {
-                x: res.x,
-                leftmost: res.leftmost,
-                rightmost: res.rightmost,
-                childrens: [res],
-            };
-        }
         let width = BlockConst.UNIT_WIDTH;
         let leftmost = 0.0;
         let rightmost = 0.0;
+        let bottommost = 0.0;
+        let bottomdiff = 0.0;
         let childrens: FormatResult[] = [];
         let i = 0;
         for (const child of children) {
@@ -185,6 +183,8 @@ export class DefineBlock extends ParentBlock {
             } else {
                 width += res.rightmost - res.leftmost;
             }
+            bottommost = Math.max(bottommost, res.bottommost);
+            bottomdiff = Math.max(bottomdiff, res.bottomdiff - this.get_height());
             childrens.push(res);
             i += 1;
         }
@@ -195,8 +195,10 @@ export class DefineBlock extends ParentBlock {
             + this.get_width() * 0.5;
         return {
             x: x,
-            leftmost: leftmost,
-            rightmost: rightmost,
+            leftmost: leftmost - x,
+            rightmost: rightmost + x,
+            bottommost: bottommost + BlockConst.UNIT_HEIGHT * 2 + BlockConst.DEFINE_BLOCK_BORDER * 2,
+            bottomdiff: bottomdiff + (bottommost + BlockConst.DEFINE_BLOCK_BORDER * 2) * 0.5,
             childrens: childrens,
         };
     }
